+++ packages/echo-base/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@equinor/echo-base",
-    "version": "0.2.96",
+    "version": "0.2.97",
     "module": "esm/index.js",
     "main": "lib/index.js",
     "typings": "lib/index.d.ts",
