+++ packages/echo-base/src/index.ts
@@ -3,5 +3,3 @@ export * from './module';
 export * from './types';
 export { EchoEvents } from './types/event';
 export * from './utils';
-export { default as eventHub } from './utils/eventHub';
-
