+++ packages/echo-core/package.json
@@ -29,7 +29,7 @@
         "directory": "packages/echo-core"
     },
     "peerDependencies": {
-        "@equinor/echo-base": "0.2.93",
+        "@equinor/echo-base": "^0.2.97",
         "react": ">=17.0.2",
         "react-dom": ">=17.0.2",
         "react-router": ">=5.2.0",
@@ -44,7 +44,7 @@
         "@babel/preset-env": "^7.12.11",
         "@babel/preset-react": "^7.12.10",
         "@babel/preset-typescript": "^7.12.7",
-        "@equinor/echo-base": "^0.2.93",
+        "@equinor/echo-base": "0.2.97",
         "@microsoft/microsoft-graph-types": "^1.20.0",
         "@rollup/plugin-node-resolve": "^11.0.1",
         "@rollup/plugin-typescript": "^8.0.0",
