+++ packages/echo-core/src/EchoCore.ts
@@ -1,0 +1,30 @@
+import * as error from '@equinor/echo-base/lib/errors';
+import { setLegendOption } from './actions/legendOptions';
+import * as moduleActions from './actions/moduleState';
+import * as hooks from './hooks';
+import { useAuthenticate } from './hooks/useAuthenticate';
+import { useLegendOptions } from './hooks/useLegendOptions';
+import { useEchoSetup } from './hooks/useSetup';
+import { useUserPhoto } from './hooks/useUserPhoto';
+import { useUserProfile } from './hooks/useUserProfile';
+import { EchoAuthProvider } from './services/authentication/echoProvider';
+import echoClient from './services/echoClient/echoClient';
+import * as moduleState from './state';
+import { useAppModuleState } from './state/useAppModuleState';
+
+export const EchoCore = {
+    useEchoSetup: useEchoSetup,
+    useAppModuleState: useAppModuleState,
+    useLegendOptions: useLegendOptions,
+    useUserProfile: useUserProfile,
+    useUserPhoto: useUserPhoto,
+    useAuthenticate: useAuthenticate,
+    setLegendOption: setLegendOption,
+    EchoAuthProvider: EchoAuthProvider,
+    EchoClient: echoClient,
+
+    // Exposing all core Hooks
+    hooks,
+    moduleState: { ...moduleState, ...moduleActions },
+    error
+};
