+++ packages/echo-core/src/index.ts
@@ -1,18 +1,6 @@
-import * as error from '@equinor/echo-base/lib/errors';
-import { setLegendOption } from './actions/legendOptions';
-import * as moduleActions from './actions/moduleState';
 import { EnvironmentVariables } from './configuration/environment';
+import { EchoCore } from './EchoCore';
 import Env from './Env';
-import * as hooks from './hooks';
-import { useAuthenticate } from './hooks/useAuthenticate';
-import { useLegendOptions } from './hooks/useLegendOptions';
-import { useEchoSetup } from './hooks/useSetup';
-import { useUserPhoto } from './hooks/useUserPhoto';
-import { useUserProfile } from './hooks/useUserProfile';
-import { EchoAuthProvider } from './services/authentication/echoProvider';
-import echoClient from './services/echoClient/echoClient';
-import * as moduleState from './state';
-import { useAppModuleState } from './state/useAppModuleState';
 
 export * from '@azure/msal-browser';
 export { EchoEvents, storage } from '@equinor/echo-base';
@@ -35,8 +23,6 @@ export { default as EchoEventHubProvider } from './providers/EchoEventHubProvide
 export { AuthenticationProvider } from './services/authentication/authProvider';
 export { default as BaseClient } from './services/baseClient/baseClient';
 export * from './settings';
-export * from './settings/plantSettingsActions';
-export * from './settings/useSetting';
 export * from './state/useAppModuleState';
 export { useAppModuleState } from './state/useAppModuleState';
 export * from './state/useGlobalState';
@@ -44,23 +30,6 @@ export * from './types';
 export { makeUniqBy } from './utils/uniq';
 
 export const EchoEnv = new Env();
-
-export const EchoCore = {
-    useEchoSetup: useEchoSetup,
-    useAppModuleState: useAppModuleState,
-    useLegendOptions: useLegendOptions,
-    useUserProfile: useUserProfile,
-    useUserPhoto: useUserPhoto,
-    useAuthenticate: useAuthenticate,
-    setLegendOption: setLegendOption,
-    EchoAuthProvider: EchoAuthProvider,
-    EchoClient: echoClient,
-
-    // Exposing all core Hooks
-    hooks,
-    moduleState: { ...moduleState, ...moduleActions },
-    error
-};
 
 export default EchoCore;
